packages/arcllm/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arcllm"
version = "0.1.0"
description = "Unified LLM abstraction layer for autonomous agents"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.25",
    "opentelemetry-api>=1.20",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
]
signing = [
    "cryptography>=42.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
arcllm = ["*.toml", "providers/*.toml"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
